Add files via upload
At this stage I was in the process of installing Apache, I ran into an a few errors on the first trial. Errors occurred due to running configuration directives (Server Root) directly in Command Prompt instead of editing httpd.conf, and attempting to execute directory paths as commands. Apache was already installed as a service, so reinstalling returned a warning. Apache failed to start because port 80 was already in use, preventing it from binding to the listening socket and causing shutdown.